### 2020-10-15 #61 @whedon
[PRE REVIEW]: DFTK: A Julian approach for simulating electrons in solids
**Submitting author:** @mfherbst (<a href="http://orcid.org/0000-0003-0378-7921">Michael F. Herbst</a>)
**Repository:** <a href="https://github.com/mfherbst/juliacon2020-proceedings-dftk.git" target ="_blank">https://github.com/mfherbst/juliacon2020-proceedings-dftk.git</a>
**Version:**
