### 2021-01-30 #61 @crstnbr
Hi @mfherbst, you're confusion about why this isn't working is understandable. Our website apparently contains outdated information which I only now am in the process of fixing.

Anyways, the issue is that in contrast to JOSS we are accepting LaTeX only submissions. So instead of a paper.md etc. you should have a `paper.tex` file. More specifically, check out our new [author's guide](https://juliacon.github.io/proceedings-guide/author/). It points you to a template repository that shows you exactly how your repository/`paper/` subfolder should look like. It should be rather straightforward to port your paper repo to this format.

Sorry for the inconvenience!

BTW: If you think there is something unclear in/to be added to the author's guide, I'd very much appreciate some feedback!
