+++ paper/paper.tex
@@ -1,27 +1,11 @@
----
-title: 'DFTK: A Julian approach for simulating electrons in solids'
-tags:
-  - Julia
-  - density-functional theory
-  - electronic structure theory
-  - numerical analysis
-  - solid state
-  - ab initio
+\documentclass{juliacon}
+\setcounter{page}{1}
 
-authors:
-  - name: Michael F. Herbst^[Corresponding author]
-    orcid: 0000-0003-0378-7921
-    affiliation: 1
-  - name: Antoine Levitt
-    affiliation: 1
-  - name: Eric Cancès
-    affiliation: 1
-affiliations:
- - name: CERMICS, Ecole des Ponts and Inria Paris, 6 & 8 avenue Blaise Pascal, 77455 Marne-la-Vallée, France
-   index: 1
-bibliography: biblio.bib
-date: 28 September 2020
----
+\begin{document}
+
+\input{header}
+
+\maketitle
 
 Density-functional theory (DFT) is a widespread method for simulating the
 quantum-chemical behaviour of electrons in matter.
@@ -32,15 +16,17 @@ Tackling such questions successfully often requires
 multidisciplinary collaboration,
 where a difficulty is to be able to bundle the
 research efforts of all involved communities in a single software platform.
-With the density-functional toolkit (DFTK, https://dftk.org)
+With the density-functional toolkit (DFTK, \url{https://dftk.org})
 we provide a Julia package for DFT simulations
-in solid-state systems to support such interdisciplinary research.
+in solid-state systems to support such interdisciplinary research,
+see Figure~\ref{fig:logodftk}.
 With an accessible code base (around 5000 lines)
 and a performance comparable to established DFT packages
 DFTK is both fast enough for practical calculations,
 but also offers the flexibility to support fundamental mathematical research
 on reduced models and toy problems.
-See https://docs.dftk.org for documentation and examples to get started.
+See \url{https://docs.dftk.org} for documentation and examples to get started.
+\vskip 6pt
 
 A recent approach for DFT-based simulations
 in chemistry and materials science are high-throughput screening studies.
@@ -59,6 +45,7 @@ to support upcoming hardware architectures,
 unusual computational workflows
 or fundamental algorithm development
 based on model problems.
+\vskip 6pt
 
 With DFTK we want to overcome these obstacles and in particular
 facilitate the collaboration
@@ -74,21 +61,38 @@ and features a low entrance barrier for new researchers
 by keeping the structure of the code in line with the
 mathematical and physical structure of DFT.
 As a result DFTK is highly suitable to rapidly prototype
-new physical models or support the mathematical analysis of DFT methods.
-In recent works we used DFTK to
-(a) derive *a posteriori* error estimates for simple DFT-like models,
+new physical models or support the mathematical analysis of DFT methods,
+which has been demonstrated in a number
+of recent papers~\cite{scfanalysis,garrigue2021building,scfprecondition,scferror}
+by us and our collaborators.
+In particular our recent research employed DFTK
+(a) to derive \textit{a posteriori} error estimates for simple DFT-like models~\cite{scferror},
 which will help to construct automatic error balancing strategies,
-and (b) develop black-box preconditioning strategies
-for calculations on large inhomogeneous systems
+and (b) to develop black-box preconditioning strategies
+for calculations on large inhomogeneous systems~\cite{scfprecondition}
 based on a parameter-free physical model.
 Both these projects were only possible since DFTK allowed us to tackle
 reduced models, where one could obtain mathematical or physical insight,
 and then test the new methods on realistic full-scale applications.
 We hope that DFTK will be a useful platform
 for future multidisciplinary developments in density-functional theory.
+\vskip 6pt
 
[…]
