+++ paper/ref.bib
@@ -1,0 +1,42 @@
+@article{scferror,
+  author = {Herbst, Michael F. and Levitt, Antoine and Canc\`{e}s, Eric},
+  doi = {10.1039/d0fd00048e},
+  journal = {Faraday Discuss.},
+  language = {en},
+  publisher = {Royal Society of Chemistry (RSC)},
+  title = {A posteriori error estimation for the non-self-consistent Kohn-Sham equations},
+  volume = {224},
+  pages = {227},
+  url = {http://dx.doi.org/10.1039/d0fd00048e},
+  year = {2020},
+}
+
+@article{scfprecondition,
+  author = {Herbst, Michael F. and Levitt, Antoine},
+  doi = {10.1088/1361-648x/abcbdb},
+  journal = {J. Phys. Condens. Matter},
+  publisher = {IOP Publishing},
+  title = {Black-box inhomogeneous preconditioning for self-consistent field iterations in density functional theory},
+  volume = {33},
+  pages = {085503},
+  url = {http://dx.doi.org/10.1088/1361-648x/abcbdb},
+  year = {2021},
+}
+
+@unpublished{scfanalysis,
+  author = {Cancès, Eric and Kemlin, Gaspard and Levitt, Antoine},
+  title = {Convergence analysis of direct minimization and self-consistent iterations},
+  url = {https://hal.inria.fr/hal-02546060},
+  note = {arXiv:2004.09088},
+  year = {Submitted},
+}
+
+@unpublished{garrigue2021building,
+      title={Building Kohn-Sham potentials for ground and excited states}, 
+      author={Louis Garrigue},
+      year={Submitted},
+      eprint={2101.01127},
+      archivePrefix={arXiv},
+      primaryClass={math-ph},
+      note = {arxiv:2101.01127},
+}
