+++ paper/header.tex
@@ -1,0 +1,18 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+
+\title{DFTK: A Julian approach for simulating electrons in solids}
+
+\author[1]{Michael F. Herbst}
+\author[1]{Antoine Levitt}
+\author[1]{Eric Cancès}
+\affil[1]{CERMICS, Ecole des Ponts and Inria Paris, 6 \& 8 Avenue Blaise Pascal, 77455 Marne-la-Vallée, France}
+
+\keywords{Julia, density-functional theory, electronic structure theory, numerical analysis, solid state, ab initio}
+
+\hypersetup{
+pdftitle = {DFTK: A Julian approach for simulating electrons in solids},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {Michael F. Herbst, Antoine Levitt, Eric Cancès},
+pdfkeywords = {Julia, density-functional theory, electronic structure theory, numerical analysis, solid state, ab initio},
+}
+
