### 2021-01-31 #61 @crstnbr
@mfherbst Great to see that compilation worked this time!

Typically our standard for the "extended abstract" papers is one page **excluding** references. Do you think it is possible to slightly extend the current paper in a meaningful way to meet this criterium?

Also it would be great if you could suggest potential referees. Physics + Julia background would be ideal but is no strict requirement.
