### 2021-01-31 #61 @mfherbst
@crstnbr Sure I can extend the paper a little to meet that criterion, but I can't guarantee I get to it within the next two weeks as my schedule is pretty heavy.

Regarding referees in the intersection of quantum physics and Julia, how about: davpoolechem, haampie, jagot, jgreener64, PhilipVinc.
