### 2021-02-02 #61 @crstnbr
> Sure I can extend the paper a little to meet that criterion, but I can't guarantee I get to it within the next two weeks as my schedule is pretty heavy.

That's perfectly fine, just let whedon recompile the article proof once you've made the changes.
