+++ paper/paper.tex
@@ -1,5 +1,6 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
+\usepackage{lipsum}
 
 \begin{document}
 
@@ -9,73 +10,101 @@
 
 Density-functional theory (DFT) is a widespread method for simulating the
 quantum-chemical behaviour of electrons in matter.
+It provides an accurate first-principles description of
+many optical, mechanical and chemical properties
+for a large range of systems at an acceptable computational cost.
 The desire to tackle even bigger systems and
-more involved materials, however, keeps posing novel challenges with respect to
-reliability and performance of methods such as DFT.
-Tackling such questions successfully often requires
-multidisciplinary collaboration,
-where a difficulty is to be able to bundle the
-research efforts of all involved communities in a single software platform.
-With the density-functional toolkit (DFTK, \url{https://dftk.org})
-we provide a Julia package for DFT simulations
-in solid-state systems to support such interdisciplinary research,
+more involved materials, however, keeps posing novel challenges
+that require methods to constantly improve.
+One example are so-called high-throughput screening approaches,
+which are becoming more and more prominent in recent years.
+In these techniques one
+wishes to systematically scan over huge design spaces of compounds
+in order to identify promising novel materials
+for targeted follow-up investigation.
+This has already lead to many success stories~\cite{Jain2016},
+such as the discovery of
+novel earth-abundant semiconductors~\cite{Hinuma2016},
+novel light-absorbing materials~\cite{Zhao2017},
+electrocatalysts~\cite{Greeley2006},
+materials for hydrogen storage~\cite{Hummelshoj2009}
+or for Li-ion batteries~\cite{Ahmad2018}.
+Keeping in mind the large range of physics that needs
+to be covered in these studies as well as
+the typical  number of calculations (up to the order of millions),
+a bottleneck in these studies
+is the reliability and performance
+of the underlying DFT codes.
+\vskip 6pt
+
+For tackling these aspects
+multidisciplinary collaboration is essential with
+mathematicians developing more numerically stable algorithms,
+computer scientists providing high-performance implementations,
+physicists and chemists designing appropriate models,
+and application scientists integrating the resulting methods
+inside a suitable simulation workflow.
+While to date already a sizeable number of DFT codes exist,
+e.g.~ABINIT~\cite{Romero2020}, Quantum-Espresso~\cite{Giannozzi2020}
+or VASP~\cite{VASP} to name only a few,
+they suffer from insufficient flexibility inside
+their low-level computational routines
+to easily support fundamental research
+in computer science or mathematics.
+To test emerging approaches motivated from these subjects in DFT
+---
+such as automatic differentiation, multi-precision methods,
+GPU acceleration, error estimation and numerical analysis
+---
+these communities have in the past resorted
+to developing their own tailored codes.
+These, however, are not performant enough
+to scale to the level required to test
+resulting novel DFT approaches in practical simulations.
+\vskip 6pt
+
+To overcome this barrier we have developed
+the density-functional toolkit (DFTK, \url{https://dftk.org}),
+a Julia package for DFT simulations in solid-state systems
+with the explicit goal to bundle the research efforts of all involved communities,
 see Figure~\ref{fig:logodftk}.
-With an accessible code base (around 5000 lines)
+With an accessible code base (around 6000 lines)
 and a performance comparable to established DFT packages
-DFTK is both fast enough for practical calculations,
-but also offers the flexibility to support fundamental mathematical research
-on reduced models and toy problems.
-See \url{https://docs.dftk.org} for documentation and examples to get started.
-\vskip 6pt
-
-A recent approach for DFT-based simulations
-in chemistry and materials science are high-throughput screening studies.
-In these approaches one performs systematic calculations on
[…]
