+++ paper/ref.bib
@@ -4,7 +4,7 @@
   journal = {Faraday Discuss.},
   language = {en},
   publisher = {Royal Society of Chemistry (RSC)},
-  title = {A posteriori error estimation for the non-self-consistent Kohn-Sham equations},
+  title = {{A posteriori error estimation for the non-self-consistent Kohn-Sham equations}},
   volume = {224},
   pages = {227},
   url = {http://dx.doi.org/10.1039/d0fd00048e},
@@ -32,11 +32,196 @@
 }
 
 @unpublished{garrigue2021building,
-      title={Building Kohn-Sham potentials for ground and excited states}, 
+      title={{Building Kohn-Sham potentials for ground and excited states}}, 
       author={Louis Garrigue},
       year={Submitted},
       eprint={2101.01127},
       archivePrefix={arXiv},
       primaryClass={math-ph},
       note = {arxiv:2101.01127},
+}
+
+
+@article{Jain2016,
+  author = {Jain, Anubhav and Shin, Yongwoo and Persson, Kristin A.},
+  doi = {10.1038/natrevmats.2015.4},
+  issue = {1},
+  journal = {Nat. Rev. Mater.},
+  language = {en},
+  month = {1},
+  publisher = {Springer Science and Business Media LLC},
+  title = {Computational predictions of energy materials using density functional theory},
+  type = {article},
+  url = {http://dx.doi.org/10.1038/natrevmats.2015.4},
+  volume = {1},
+  year = {2016},
+}
+
+
+@article{Hinuma2016,
+  author = {Hinuma, Yoyo and Hatakeyama, Taisuke and Kumagai, Yu and others},
+  doi = {10.1038/ncomms11962},
+  issue = {1},
+  journal = {Nat. Commun.},
+  language = {en},
+  month = {9},
+  publisher = {Springer Science and Business Media LLC},
+  title = {Discovery of earth-abundant nitride semiconductors by computational screening and high-pressure synthesis},
+  type = {article},
+  url = {http://dx.doi.org/10.1038/ncomms11962},
+  volume = {7},
+  year = {2016},
+}
+@article{Zhao2017,
+  author = {Zhao, Xin-Gang and Yang, Dongwen and Sun, Yuanhui and others},
+  doi = {10.1021/jacs.7b02120},
+  issue = {19},
+  journal = {J. Amer. Chem. Soc.},
+  language = {en},
+  month = {5},
+  pages = {6718--6725},
+  publisher = {American Chemical Society (ACS)},
+  title = {{Cu\textendash{}In Halide Perovskite Solar Absorbers}},
+  type = {article},
+  url = {http://dx.doi.org/10.1021/jacs.7b02120},
+  volume = {139},
+  year = {2017},
+}
+@article{Greeley2006,
+  author = {Greeley, Jeff and Jaramillo, Thomas F. and Bonde, Jacob and others},
+  doi = {10.1038/nmat1752},
+  issue = {11},
+  journal = {Nat. Mater.},
+  language = {en},
+  month = {11},
+  pages = {909--913},
+  publisher = {Springer Science and Business Media LLC},
+  title = {Computational high-throughput screening of electrocatalytic materials for hydrogen evolution},
+  type = {article},
+  url = {http://dx.doi.org/10.1038/nmat1752},
+  volume = {5},
+  year = {2006},
+}
+@article{Hummelshoj2009,
+  author = {Hummelsh\o{}j, J. S. and Landis, D. D. and Voss, J. and others},
+  doi = {10.1063/1.3148892},
+  issue = {1},
+  journal = {J. Chem. Phys.},
+  language = {en},
+  month = {7},
+  pages = {014101},
[…]
