+++ paper/paper.tex
@@ -1,6 +1,5 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
-\usepackage{lipsum}
 
 \begin{document}
 
