### 2021-02-07 #61 @mfherbst
Interesting ... so whedon sees the new version, because it extracts all the new references, but the PDF it builds is still the version I submitted a few weeks ago. @crstnbr any idea why?
