### 2021-02-09 #69 @whedon
[REVIEW]: DFTK: A Julian approach for simulating electrons in solids
**Submitting author:** @mfherbst (<a href="http://orcid.org/0000-0003-0378-7921">Michael F. Herbst</a>)
**Repository:** <a href="https://github.com/mfherbst/juliacon2020-proceedings-dftk.git" target ="_blank">https://github.com/mfherbst/juliacon2020-proceedings-dftk.git</a>
**Version:**

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
