### 2021-03-13 #69 @jagot
@mfherbst I would rewrite
> For tackling these aspects multidisciplinary collaboration is essen- tial with mathematicians developing more numerically stable al- gorithms, computer scientists providing high-performance imple- mentations, physicists and chemists designing appropriate models, and application scientists integrating the resulting methods inside a suitable simulation workflow.

as something like

> To tackle these aspects, multidisciplinary collaboration with mathematicians developing more numerically stable algorithms, computer scientists providing high-performance implementations, physicists and chemists designing appropriate models, and application scientists integrating the resulting methods inside a suitable simulation workflow is essential.

Also, when I install according to the instructions and try the first tutorial, I get the following warning:
```julia
┌ Warning: Package DFTK does not have Plots in its dependencies:
│ - If you have DFTK checked out for development and have
│   added Plots as a dependency but haven't updated your primary
│   environment's manifest file, try `Pkg.resolve()`.
│ - Otherwise you may need to report an issue with DFTK
└ Loading Plots into DFTK from project dependency, future warnings for DFTK are suppressed.
```

@MasonProtter Until https://github.com/JuliaMolSim/DFTK.jl/issues/418 is fixed, you have to do a little bit of monkey patching to run the tutorial:
```julia
function DFTK.pymatgen_lattice(lattice::AbstractArray)
           # Notice: Pymatgen uses rows as lattice vectors, so we unpeel
           # our lattice column by column. The default unit in pymatgen is Ångström
           mg = pyimport("pymatgen.core.lattice")
           bohr_to_A = 1 / austrip(1u"Å")
           mg.Lattice([Array(bohr_to_A .* lattice[:, i]) for i in 1:3])
       end

julia> function DFTK.pymatgen_structure(model_or_lattice, atoms)
           mg = pyimport("pymatgen.core.structure")
           pylattice = DFTK.pymatgen_lattice(model_or_lattice)

           n_species = sum(length(pos) for (spec, pos) in atoms)
           pyspecies = Vector{Int}(undef, n_species)
           pypositions = Array{Vector{Float64}}(undef, n_species)
           ispec = 1
           for (spec, pos) in atoms
               for coord in pos
                   pyspecies[ispec] = spec.Z
                   pypositions[ispec] = Vector{Float64}(coord)
                   ispec = ispec + 1
               end
           end
           @assert ispec == n_species + 1

           mg.Structure(pylattice, pyspecies, pypositions)
       end
```
