### 2021-03-13 #69 @mfherbst
@jagot Thanks for the review ... and yes, sorry about the pymatgen issue. That's the second time such a breaking interface change has caused DFTK to stop working properly unfortunately. That's why long-term we definitely want to get rid of it.

> Also, when I install according to the instructions and try the first tutorial, I get the following warning:

Thanks. That's because we load Plots using `Requires`, so in some sense it is a "fake-warning". It seems to be fixed with https://github.com/JuliaMolSim/DFTK.jl/pull/419.
