### 2021-03-13 #69 @mfherbst
Both things should be fixed in the (latest) DFTK 0.2.7 by the way.
