### 2021-04-02 #69 @MasonProtter
Okay, sorry for all the delays. I really enjoyed reading the paper and exploring the package. This is fantastic work with a slick, powerful user-experience, well written documentation and a ton of features.

A few comments:

When the checklist asks me if the software source code is available at the respository URL, it leads me to https://github.com/mfherbst/juliacon2020-proceedings-dftk rather than https://github.com/JuliaMolSim/DFTK.jl

I would echo @jagot in his suggestion for tweaking the paragraph starting with "For tacking these aspects [...]"

Other than these tiny issues, as someone who is familiar with Monte Carlo and DMRG methods, but not so familiar with DFT methods, I find myself wanting at least a couple sentences either in the documentation or (if possible) in the paper which gives me a bit more context how I should expect this general methodolgy to compare to different techniques. Are there any rough rules of thumb I can use for when I should expect DFT to give inaccurate results relative to 'reality'?

For instance, I have certainly heard in other places that DFT methods tend to not properly model exchange interactions. I see in the examples section discussion of magnetic systems. I'm curious how well do these calculations agree with experiments, monte carlo, or other benchmarks?

I don't need anything super in-depth to give my full approval, but I think it would help me in placing this work in the wider scientific context.
