### 2021-04-02 #69 @MasonProtter
I should also say, if you know of a good reference which discusses the limitations and domains of applicability of these sorts of methods, and you feel that the comments in reference is straightforwardly applicable to your work, then that would also satisfy me!
