+++ paper/ref.bib
@@ -225,3 +225,32 @@
     year = 2021,
     howpublished = {\url{https://vasp.at}},
 }
+@article{Burke2012,
+  author = {Burke, Kieron},
+  doi = {10.1063/1.4704546},
+  issue = {15},
+  journal = {J. Chem. Phys.},
+  month = {4},
+  pages = {150901},
+  title = {Perspective on density functional theory},
+  volume = {136},
+  year = {2012},
+}
+@article{Burke2013,
+  author = {Burke, Kieron and Wagner, Lucas O.},
+  doi = {10.1002/qua.24259},
+  journal = {Int. J. Quantum Chem.},
+  number = {2},
+  pages = {96-101},
+  title = {DFT in a nutshell},
+  volume = {113},
+  year = {2013},
+}
+@book{Martin2004,
+  author = {Martin, Richard M.},
+  doi = {10.1017/cbo9780511805769},
+  isbn = {9780511805769},
+  publisher = {Cambridge University Press},
+  title = {Electronic Structure},
+  year = {2004},
+}
