### 2021-04-06 #69 @mfherbst
> For instance, I have certainly heard in other places that DFT methods tend to not properly model exchange interactions. I see in the examples section discussion of magnetic systems. I'm curious how well do these calculations agree with experiments, monte carlo, or other benchmarks?

Thanks for your review and for this guiding question. The problem with answering it, is that it's really tough to do that in just a couple of sentences and at the same time convey a fair picture. We decided to not go into this, simply because this is more of a software paper and not an application paper after all.

But I agree, that we don't really point to any literature answering this question either, which we should. I amended the paper appropriately and [made a PR to the repo](https://github.com/JuliaMolSim/DFTK.jl/pull/422). @MasonProtter let me know if that is not sufficient from your point of view.
