### 2021-04-06 #69 @MasonProtter
Yes, this is great.

One other tiny detail: I don't know what the convention is for the JuliaCon proceedings, but in the linked PDF proof, the references are not in the order in which they appear in the paper. I don't have a super strong opinion on it, but it's maybe worth tweaking.

I'm ready to sign off my approval though. :)
