### 2021-04-07 #69 @mfherbst
@MasonProtter Thanks for your feedback :+1:.

>  the references are not in the order in which they appear in the paper.

I'm not a fan of this either, but it's the same in other JuliaCon papers, so I assume that is the style to be followed.
