### 2021-04-08 #69 @ranjanan
>the references are not in the order in which they appear in the paper.

Just to clarify: this is the style of references that we currently use. cc: @vchuravy @matbesancon @crstnbr
