### 2021-04-22 #69 @crstnbr
@mfherbst Since both reviewers signed off their approval, we're almost done here!

The only thing that seems a bit odd to me is that the first heading is "1. Acknowledgments". Can you try to add a heading at the beginning (before the main text), for example "1. Description" or similar, such that acknowledgments becomes the second heading.

(I see that this might push parts of the acknowledgments section to the second page. Perhaps we should move the entire section to the second page. But let's just see how it looks.)
