### 2021-04-22 #69 @mfherbst
I'm not sure it makes a lot of sense to add such a "fake heading". How about making the Acknowledgements a `\section*` such that the `1. ` is missing? Would that be ok with the Journal style?
