### 2021-04-22 #69 @crstnbr
Yes, that would work as well! But please also drop the "2." from "2. References". Thanks.
