+++ paper/paper.tex
@@ -118,7 +118,8 @@ See \url{https://docs.dftk.org} for documentation and examples to get started wi
     \label{fig:logodftk}
 \end{figure}
 
-\section{Acknowledgements} This project has received funding from the Institute of
+\section*{Acknowledgements}
+This project has received funding from the Institute of
 Computing and Data Sciences (ISCD, Sorbonne Université), École des Ponts
 ParisTech, Inria Research Centre Paris and from the European Research Council
 (ERC) under the European Union's Horizon 2020 research and innovation program
