### 2021-04-22 #69 @mfherbst
@crstnbr That seems to have worked: Acknowledgements and references without section numbers.
