### 2021-04-23 #69 @crstnbr
@mfherbst Can you confirm that 10.1137/20m1332864 is a valid DOI for the "Convergence analysis of direct minimization and self-consistent iterations" reference? (It looks like it.)

If so, please add the doi to the corresponding bib entry. Thanks!
