+++ paper/ref.bib
@@ -23,12 +23,16 @@
   year = {2021},
 }
 
-@unpublished{scfanalysis,
+@article{scfanalysis,
   author = {Cancès, Eric and Kemlin, Gaspard and Levitt, Antoine},
-  title = {Convergence analysis of direct minimization and self-consistent iterations},
-  url = {https://hal.inria.fr/hal-02546060},
-  note = {arXiv:2004.09088},
-  year = {Submitted},
+  doi = {10.1137/20M1332864},
+  journal = {SIAM J. Matrix Anal. Appl.},
+  volume = {42},
+  number = {1},
+  pages = {243-274},
+  year = {2021},
+  title = {Convergence Analysis of Direct Minimization and Self-Consistent Iterations},
+  doi = {10.1137/20M1332864},
 }
 
 @unpublished{garrigue2021building,
