### 2021-04-23 #69 @crstnbr
@mfherbst Final thing, can you please archive the current state of the DFTK repository, for example using Zenodo and post the associated DOI here (see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance).

Afterwards I can tell whedon to accept the paper!
